Remove requirement for super with ints module functions
Also improve set_ipl() documentation.

Additionally, the function prototypes have changed to use short instead
of int. This will help the functions work the same across different
compilers with differently defined definitions for int.

+++ src/ints.h
@@ -14,16 +14,20 @@
 
 #define MASK_ALL_INTERRUPTS 07
 
-/* sets IPL = mask, returns old IPL (interrupt priority level) */
-int set_ipl(int mask); /* must be called from supervisor mode */
+/**
+ * @brief Set the CPU's interrupt mask to the given value.
+ * 
+ * @param mask The Interrupt Priority Level to set.
+ * @return The old Interrupt Priority Level.
+ */
+short set_ipl(short mask);
 
 /**
  * @brief Toggles masking for a specific device with the given priority level in
  * the MC68901.
  * 
  * @param channel A four bit value that specifies the device to be masked.
- * @note Requires super privileges.
  */
-void mask_level_toggle(int channel);
+void mask_level_toggle(short channel);
 
 #endif
